Use replaygain tag, if it exists
If the input file contains a replay-gain tag, it will be used in the ffmpeg conversion to the raw file to send to the player. This will make it possible to use metaflac, for example, to normalize song levels if taken from different input levels. If the tag doesn't exist, nothing will be done.